requirements.txt
pandas
scipy
wandb
torch
torchvision
python-dotenv